Initialize the configuration part
There are likely to be more options, data types to be researched
TODO
Parsing of the config file into the pishutter script
Currently planning to reload config at startup.
Possibly we should allow a reload config mechanical selector,
or perhaps a scheduled software reload.
The constraint to keep in mind is instantenous image capture on button press
Reloading the conf file from disk is going to be slow,
however that needs to be measured. Maybe it doesnt matter,
and we can just reload conf file at button press